Applying Aggregate 01
Please note that this change is for school project on Applying Aggregates. So please bear with me.

The problem that I noticed is that there is no clear design to make sure that the design follows a consistent valid state for each iteration. Therefore I implemented an aggregate to make sure that run.py follows a valid state before moving on to the caller methods