Apply Singleton Design Pattern to the Person Class
Applying Aggregate Design Pattern to population.py

Background
This pull request is part of a course project I am doing for EECS3311 at York University. We are tasked with applying design patterns we have learned to covasim to improve the functionality of the application in a way that demonstrates our knowledge of the course material.

Abstract
I discovered a deprecated class within Covasim that permitted multiple instances, which is a potential security flaw and vector for memory loss as well as not being of any particular use to the application at large.

Explanation
I refactored the code to make the Person class align with the Singleton design principal, which removed the possibility of users creating more than one instance of it. This removed the security and memory issues, and also paves the way for new functionality requiring a single instance of an agent.