Creational Design pattern
##Intro

We use Factory pattern to create an object without exposing the creation logic to the client and refer to newly
 created object using a common interface.  The factory method lets a class defer instantiates to subclasses. It lets the subclasses to decide which class to instantiate.

##Problem:

There are many methods in the plotting.py that keeps repeating its iterations when a simulation is called. By using the factory design pattern, we can create a new plotting factory class to create the objects that the simulation calls. This means that instead of calling new methods for Create_figs, Create_subplots, title_grid_legend , we can group them and combine them into a general method called plot.

##Solution/Changes:

Now we can call for this factory class, so that the client can simply run the plotting method that is combine together from the factory class, instead of the simulation running each method separately on the plotting file.